refactored route settings class into separate file
